Move reimplementation to separate module for clarity
